Various bug fixed and changes
* Adjust demo curricula
* Fix training buffer reset bug
* Make wall height a float
* Add pertained models for Area env

+++ unity-environment/Assets/ML-Agents/Examples/Area/Scripts/Wall/WallAcademy.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 public class WallAcademy : Academy {
 
-    public int minWallHeight;
-    public int maxWallHeight;
+    public float minWallHeight;
+    public float maxWallHeight;
 
 	public override void AcademyReset()
 	{
-        minWallHeight = (int)resetParameters["min_wall_height"];
-        maxWallHeight = (int)resetParameters["max_wall_height"];
+        minWallHeight = (float)resetParameters["min_wall_height"];
+        maxWallHeight = (float)resetParameters["max_wall_height"];
 	}
 
 	public override void AcademyStep()
